+++ packages/react-server/examples/next/app/actions/client/page.js
@@ -1,13 +1,12 @@
 "use client";
 
-import "client-only";
 import { useState } from "react";
 
 import double, {
   inc,
   dec,
   redirectAction,
-  // getHeaders,
+  getHeaders,
   renamed,
   slowInc,
 } from "./actions";
@@ -58,15 +57,18 @@ export default function Counter() {
         *2
       </button>
       <form>
-        <button id="redirect-pages" formAction={() => redirectAction("/")}>
-          redirect
+        <button
+          id="redirect-pages"
+          formAction={() => redirectAction("/pages-dir")}
+        >
+          redirect to a pages route
         </button>
       </form>
-      {/* <form action={getHeaders}>
+      <form action={getHeaders}>
         <button type="submit" id="get-header">
           submit
         </button>
-      </form> */}
+      </form>
       {/* <form action={test}>
         <button>test</button>
       </form> */}
